+++ src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "./tui/Select.js";
 export * from "./filesystem.js";
 export * from "./formatting.js";
 export * from "./logger.js";